- sql debug...
git-svn-id: http://svn.codergrid.de/svn/sf/bruteripper/trunk@887 7add9c9f-332d-0410-adcc-f7b1da7091d4

+++ src/org/esb/av/StreamInfo.cpp
@@ -76,7 +76,7 @@ namespace org {
       int StreamInfo::getDuration(){
         return _s->duration;
       }
-      int StreamInfo::getQuality(){
+      float StreamInfo::getQuality(){
         return _s->quality;
       }
       long long int StreamInfo::getFirstPts(){
