+++ src/org/esb/av/StreamInfo.h
@@ -30,7 +30,7 @@ namespace org {
         int getPixelFormat();
         int getSampleFormat();
         int getDuration();
-        int getQuality();
+        float getQuality();
         long long int getFirstPts();
         long long int getFirstDts();
       private:
